session: return ReturnCode from reset() on failure
Callers of reset() will care about a ReturnCode::NetError, since this is
what is actually typically returned on success. The HSM reboots
immediately when reset() is called, meaning it vanishes out from under
the connector which dutifully returns a NetError. However, don't try to
be clever here in the library and let callers deal with it.